fix: adds note for copilot docs
If you have a file open, when running `codegate version`, you get a response based on the file and not the codegate version itself. I've added a note to inform users that they close all files before running the command.
Outside of this PR, we may want to decide if this is desired functionality by Copilot or Codegate and if the version should be returned regardless if a file is currently open or not.